tok/requirements.txt
transformers== 4.40.1
fastembed==0.3.4
flaskwebgui==1.1.3
flask==3.0.0
flask_cors==4.0.0
llama-index-vector-stores-neo4jvector==0.1.4
llama-index-embeddings-fastembed==0.1.7
llama-index-llms-ollama==0.2.2
llama-index-readers-file==0.1.20
llama-index-core==0.10.34